use f32 instead of f64

+++ src/camera.rs
@@ -11,19 +11,19 @@ pub struct Camera {
     lower_left_corner: Point3,
     u: Vec3,
     v: Vec3,
-    lens_radius: f64
+    lens_radius: f32
 }
 
 impl Camera {
     pub fn new(
         look_from: Point3, look_at: Point3, vup: Vec3,
-        vfov: f64, aspect_ratio: f64, aperture: f64,
-        focus_dist: f64
+        vfov: f32, aspect_ratio: f32, aperture: f32,
+        focus_dist: f32
     ) -> Camera {
-        let theta: f64 = degrees_to_radians(vfov);
-        let h: f64 = (theta / 2.0).tan();
-        let viewport_height: f64 = 2.0 * h;
-        let viewport_width: f64 = aspect_ratio * viewport_height;
+        let theta: f32 = degrees_to_radians(vfov);
+        let h: f32 = (theta / 2.0).tan();
+        let viewport_height: f32 = 2.0 * h;
+        let viewport_width: f32 = aspect_ratio * viewport_height;
 
         let w = unit_vector(look_from - look_at);
         let u = unit_vector(cross(vup, w));
@@ -46,7 +46,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, s: f64, t: f64) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd: Vec3 = self.lens_radius * random_in_unit_disk();
         let offset: Vec3 = self.u * rd.x() + self.v * rd.y();
